Changed search property to int

+++ FluentVault/Requests/Search/Files/SearchFilesRequest.cs
@@ -9,7 +9,7 @@ internal class SearchFilesRequest : SessionRequest, ISearchFilesRequestBuilder, 
     private object _searchValue = new();
     private SearchOperator _operator = SearchOperator.Contains;
     private SearchPropertyType _propertyType = SearchPropertyType.SingleProperty;
-    private long _property;
+    private int _property;
 
     public SearchFilesRequest(VaultSession session)
         : base(session, RequestData.FindFilesBySearchConditions) { }
@@ -30,7 +30,7 @@ internal class SearchFilesRequest : SessionRequest, ISearchFilesRequestBuilder, 
 
     public ISearchFilesAddSearchCondition InProperty(SearchStringProperty property)
     {
-        _property = (long)property;
+        _property = (int)property;
         return this;
     }
 
@@ -121,6 +121,6 @@ internal class SearchFilesRequest : SessionRequest, ISearchFilesRequestBuilder, 
         return bodyBuilder.ToString();
     }
 
-    private static string GetSearchCondition(string searchText, long propertyId, SearchOperator searchOperator, SearchPropertyType propertyType, SearchRule searchRule)
+    private static string GetSearchCondition(string searchText, int propertyId, SearchOperator searchOperator, SearchPropertyType propertyType, SearchRule searchRule)
         => $@"<SrchCond PropDefId=""{propertyId}"" SrchOper=""{searchOperator}"" SrchTxt=""{searchText}"" PropTyp=""{propertyType}"" SrchRule=""{searchRule}""/>";
 }
